pyproject.toml
[tool.poetry]
name = "mosaic"
version = "0.1.0"
description = ""
authors = ["Mohammed <mohammed.mm404@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-wtf = "^1.2.1"
flask-login = "^0.6.3"
email-validator = "^2.2.0"
pyjwt = "^2.8.0"
flask-mail = "^0.10.0"
markdown = "^3.6"
bleach = "^6.1.0"


[tool.poetry.group.dev.dependencies]
faker = "^26.0.0"
aiosmtpd = "^1.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
